requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
streamlit>=1.28.0
joblib>=1.2.0
pyyaml>=6.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

